feat: forcing order of categorical features in OneHot encoding
sklearn-pandas's mapper transforms variables in the order that they are given to it.
Moreover, predictsignauxfaibles.explain.explain takes as input conf.FEATURES
to understand the order in which categorical features were treated.
Therefore, we must enforce in the model configuration file that categorical features
in TO_ONEHOT_ENCODE are listed in the same order as they are in conf.FEATURES